+++ UIComponents.Generators/Validators/UICValidatorReadonlyAttribute.cs
@@ -25,6 +25,11 @@ namespace UIComponents.Generators.Validators
         public async Task<bool> IsReadonly(PropertyInfo propertyInfo, object obj)
         {
             await Task.Delay(0);
+            if (!propertyInfo.CanRead)
+            {
+                _logger.LogDebug("{0} is readonly because it cannot be set", $"{propertyInfo.DeclaringType.Name}.{propertyInfo.Name}");
+                return true;
+            }
             var readonlyAttr = propertyInfo.GetCustomAttribute<ReadOnlyAttribute>();
             if (readonlyAttr != null)
             {